Checks for Robot RSDK 1.1 and newer for SSH update
This adds a check to prevent RSDK customers from attempting to update
their robot with the update_robot.py script when running robot software
that is version 1.1 or newer. This would have errored out anyway, as the
new way to update the robot is on-robot (via ruser) with the command
rethink-updater. See http://sdk.rethinkrobotics.com/wiki/SSH_Update for
more detailed instructions.